Camera: Make model into an abstract class
