+++ src/Camera/CameraModel.h
@@ -9,9 +9,12 @@
 
 #pragma once
 
-#include <chrono>
+#include <cstdint>
 
-#include <glm/vec2.hpp>
+#include <chrono>
+#include <memory>
+#include <optional>
+
 #include <glm/vec3.hpp>
 
 namespace openblack
@@ -22,53 +25,26 @@ class Camera;
 class CameraModel
 {
 public:
-	CameraModel();
-	virtual ~CameraModel() = default;
+	enum class Model : uint8_t
+	{
+		Old,
+	};
 
-	void FlyInit();
-	void ResetVelocities();
-	void Update(std::chrono::microseconds dt, const Camera& camera);
-	void HandleActions(const Camera& camera);
+	struct CameraInterpolationUpdateInfo
+	{
+		glm::vec3 origin;
+		glm::vec3 focus;
+	};
 
-	glm::vec3 GetTargetPosition() const;
-	glm::vec3 GetTargetFocus() const;
+	static std::unique_ptr<CameraModel> CreateModel(Model model);
 
-protected:
-	glm::vec3 _targetPosition;
-	glm::vec3 _targetRotation;
-	glm::vec3 _dv;
-	glm::vec3 _dwv;
-	glm::vec3 _dsv;
-	glm::vec3 _ddv;
-	glm::vec3 _duv;
-	glm::vec3 _drv;
-	glm::vec3 _velocity;
-	glm::vec3 _hVelocity;
-	glm::vec3 _rotVelocity;
-	float _accelFactor;
-	float _movementSpeed;
-	float _maxMovementSpeed;
-	float _maxRotationSpeed;
-	bool _lmouseIsDown;
-	bool _mmouseIsDown;
-	bool _mouseIsMoving;
-	glm::ivec2 _mouseFirstClick;
-	bool _shiftHeld;
-	glm::ivec2 _handScreenVec;
-	float _handDragMult;
-	bool _flyInProgress;
-	float _flyDist;
-	float _flySpeed;
-	float _flyStartAngle;
-	float _flyEndAngle;
-	float _flyThreshold;
-	float _flyProgress;
-	glm::vec3 _flyFromPos;
-	glm::vec3 _flyToNorm;
-	glm::vec3 _flyFromTan;
-	glm::vec3 _flyToPos;
-	glm::vec3 _flyToTan;
-	glm::vec3 _flyPrevPos;
+	virtual ~CameraModel();
+
+	virtual std::optional<CameraInterpolationUpdateInfo> Update(std::chrono::microseconds dt, const Camera& camera) = 0;
+	virtual void HandleActions(std::chrono::microseconds dt) = 0;
+	[[nodiscard]] virtual glm::vec3 GetTargetPosition() const = 0;
+	[[nodiscard]] virtual glm::vec3 GetTargetFocus() const = 0;
+	[[nodiscard]] virtual std::chrono::seconds GetIdleTime() const = 0;
 };
 
 } // namespace openblack
